requirements.txt
# Core dependencies
python>=3.8
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# GUI and Visualization
tkinter>=8.6  # Usually comes with Python
matplotlib>=3.5.0
networkx>=2.6.3
Pillow>=9.0.0  # For image handling in Tkinter

# Data processing
scipy>=1.7.0  # For hierarchical clustering
python-dateutil>=2.8.0  # Often required by pandas

# Development tools (optional)
pytest>=6.0.0
black>=22.0.0  # Code formatter